requirements.txt
Django==4.2.23
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
drf-yasg==1.21.10
django-filter==23.5